feat:添加peer状态监听回调

+++ device/allowedips_test.go
@@ -15,7 +15,7 @@ import (
 type testPairCommonBits struct {
 	s1    []byte
 	s2    []byte
-	match uint8
+	match uint
 }
 
 func TestCommonBits(t *testing.T) {
